[FIX] : 🔧 videoAdapter 잔잔바리 수정 완료
