+++ app/src/main/java/com/example/team99/Home/CategoryAdapter.kt
@@ -1,12 +1,16 @@
 package com.example.team99.Home
 
 import android.content.Context
+import android.content.Intent
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.example.team99.MainActivity
+import com.example.team99.VideoDetailActivity
 import com.example.team99.databinding.VideoItemBinding
 
 class CategoryAdapter (private val mContext: Context) : RecyclerView.Adapter<CategoryAdapter.CategoryHolder>() {
@@ -17,6 +21,21 @@ class CategoryAdapter (private val mContext: Context) : RecyclerView.Adapter<Cat
         var thumbnails: ImageView = binding.thumbnailsIv
         var title: TextView = binding.titleTv
 
+        init {
+            thumbnails.setOnClickListener {
+                val position = adapterPosition.takeIf { it != RecyclerView.NO_POSITION } ?: return@setOnClickListener
+                val clickItem = categoryVideoItems[position]
+                if (position != RecyclerView.NO_POSITION) {
+                    MainActivity.saveSelectedItem(mContext,clickItem)
+                    Log.d("itemd", "$clickItem")
+
+                    val intent = Intent(thumbnails.context, VideoDetailActivity::class.java)
+                    intent.putExtra("title", clickItem.title)
+                    intent.putExtra("description", clickItem.description)
+                    thumbnails.context.startActivity(intent)
+                }
+            }
+        }
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CategoryHolder {
